项目/2组_基于Django的课程学习的web应用/Django-MxOnlineimooc/requirements.txt
appdirs==1.4.3
Django==1.10.5
django-crispy-forms==1.6.1
django-formtools==2.0
django-pure-pagination==0.3.0
django-simple-captcha==0.4.6
httplib2==0.9.2
mysqlclient==1.3.10
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.2.0
six==1.10.0
